Norm: move more function
