+++ headers/ft_parsing.h
@@ -6,7 +6,7 @@
 /*   By: tlassere <tlassere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/27 17:21:58 by tlassere          #+#    #+#             */
-/*   Updated: 2025/05/11 13:28:44 by tlassere         ###   ########.fr       */
+/*   Updated: 2025/05/11 13:39:12 by tlassere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,5 +78,10 @@ char	**the_mastermind(char *token_split);
 int		ft_quotes_remove(char *str);
 
 t_list	*ft_word_lst_make_join(char *str, int type);
+void ft_remove_caracter(char *str);
+char *ft_research_token(char *token, t_data *data);
+void	join_add_node(t_list **begin, t_list *add);
+int	join_create_set_node(t_list **begin, char *buffer, int type);
+char	*join_expand(t_list **lst, t_teraform *teraform, char *str);
 
 #endif
