+++ sources/ft_parsing/ft_expansion/ft_teraform.c
@@ -38,13 +38,6 @@ static int	ft_set_terraform(t_teraform *teraform, char *begin_str, char *str)
 	return (SUCCESS);
 }
 
-static char *ft_research_token(char *token, t_data *data)
-{
-	if (token[1] == '?')
-		return (ft_itoa(data->exit_program));
-	return (ft_env_tab_get_null(data->env->envp, token + 1));
-}
-
 static int	ft_set_expand(t_teraform *teraform, t_data *data)
 {
 	char	*buffer;
@@ -82,141 +75,6 @@ static void	ft_free_teraform(t_teraform *teraform)
 		free(teraform->token);
 	if (teraform->sp)
 		ft_tab_free(teraform->sp);
-}
-
-void	join_add_node(t_list **begin, t_list *add)
-{
-	if (add)
-	{
-		if (*begin)
-			ft_lstlast(*begin)->next = add;
-		else
-			*begin = add;
-	}
-}
-
-int	join_create_set_node(t_list **begin, char *buffer, int type)
-{
-	t_list	*newnode;
-
-	if (buffer == NULL)
-		return (FAIL);
-	newnode = ft_word_lst_make_join(buffer, type);
-	if (newnode == NULL)
-	{
-		free(buffer);
-		return (FAIL);
-	}
-	join_add_node(begin, newnode);
-	return (SUCCESS);
-}
-
-int	join_expand_left_node(t_list **begin, t_teraform *teraform, int type)
-{
-	char	*buffer;
-
-	if (teraform->left)
-	{
-		if (teraform->sp[0] && teraform->ljoin)
-			buffer = ft_strjoin(teraform->left, teraform->sp[0]);
-		else
-			buffer = ft_strdup(teraform->left);
-		return (join_create_set_node(begin, buffer, type));
-	}
-	return (SUCCESS);
-}
-
-int	join_expand_right_node(t_list **begin, t_teraform *teraform, int type, char **newpos)
-{
-	char	*buffer;
-	t_word	*last;
-	size_t	len;
-
-	if (teraform->right)
-	{
-		if (teraform->rjoin && *begin)
-		{
-			last = ft_lstlast(*begin)->content;
-			len = ft_strlen(last->word);
-			buffer = ft_strjoin(last->word, teraform->right);
-			if (buffer == NULL)
-				return (FAIL);
-			free(last->word);
-			last->word = buffer;
-			*newpos = buffer + len;
-		}
-		else
-		{
-			if (join_create_set_node(begin, ft_strdup(teraform->right), type))	
-				return (FAIL);
-			*newpos = ((t_word *)ft_lstlast(*begin)->content)->word;
-		}
-	}
-	return (SUCCESS);
-}
-
-int	join_middle(t_list **begin, t_teraform *teraform, int type)
-{
-	size_t	i;
-
-	i = 0;
-	if (teraform->sp[i])
-	{
-		if (teraform->left && teraform->ljoin)
-			i++;
-		while (teraform->sp[i])
-		{
-			if (join_create_set_node(begin, ft_strdup(teraform->sp[i]), type) == FAIL)
-				return (FAIL);
-			i++;
-		}
-	}
-	return (SUCCESS);
-}
-
-char	*join_nodes(t_list **lst, t_list *newlst, char *newpos)
-{
-	t_word	*buffer;
-	t_list	*last;
-
-	if (newlst == NULL)
-	{
-		buffer = (*lst)->content;
-		free(buffer->word);
-		buffer->word = NULL;
-		return (NULL);
-	}
-	last = ft_lstlast(newlst);
-	ft_word_free((*lst)->content);
-	(*lst)->content = newlst->content;	
-	if (last != newlst)
-	{
-		last->next = (*lst)->next;
-		(*lst)->next = newlst->next;
-		*lst = last;
-	}
-	free(newlst);
-	return (newpos);
-}
-
-char	*join_expand(t_list **lst, t_teraform *teraform, char *str)
-{
-	t_list	*newlst;
-	int		type;
-	char	*newpos;
-
-	newlst = NULL;
-	type = ((t_word *)(*lst)->content)->type;
-	newpos = NULL;
-	if (join_expand_left_node(&newlst, teraform, type) == FAIL)
-		return (str);
-	if (join_middle(&newlst, teraform, type) == FAIL)
-		return (ft_words_clear(&newlst), str);
-	if (join_expand_right_node(&newlst, teraform, type, &newpos) == FAIL)
-		return (ft_words_clear(&newlst), str);
-	if (newlst == NULL)
-		return (NULL);
-	return (join_nodes(lst, newlst, newpos));
 }
 
 char	*ft_expend_teraform(t_data *data, t_list **lst, char *str, int inquote)
